registration update and deleted unwanted file
