--- service_provider.kv
@@ -1,158 +1,0 @@
-<ServiceProvider>:
-    name:'service_provider'
-    BoxLayout:
-        orientation: 'vertical'
-        MDIconButton:
-            icon: "arrow-left"
-            pos_hint: {"center_y": .95}
-            user_font_size: "30sp"
-            theme_text_color: "Custom"
-            text_color: rgba(26, 24, 58, 255)
-            on_release:
-                root.manager.transition.direction = "right"
-                root.manager.current = "signup"
-        ScrollView:
-            GridLayout:
-                id: grid
-                cols: 1
-                spacing: dp(10)
-                padding: dp(20)
-                size_hint_y: None
-                height: self.minimum_height
-                BoxLayout:
-                    orientation: 'vertical'
-                    size_hint_y: None
-                    height: dp(185)
-                    md_bg_color: 1, 0, 0, 1
-
-                    Button:
-                        id: hospital_button
-                        size_hint: None, None
-                        size: (dp(290), dp(150))
-                        pos_hint: {'center_x': 0.5, 'center_y': 0.5}
-                        elevation: 0
-                        md_bg_color: 1, 1, 1, 1
-                        on_release:root.animate_button('hospital_button')
-                        Image:
-                            source: "images/hospital.png"
-                            size_hint: None, None
-                            size: dp(290), dp(150)
-                            allow_stretch: True
-                            keep_ratio: False
-                            pos:self.parent.pos
-                            md_bg_color: 1, 1, 1, 1
-                            padding: dp(10)
-                            size_hint_min: dp(290), dp(150)
-
-                    MDLabel:
-                        text: 'Hospital Facility Manager'  # Description
-                        size_hint: None, None
-                        size: dp(290), dp(35)
-                        pos_hint: {'center_x': .5}
-                        font_size: "18sp"
-                        font_name: "Roboto"
-                        halign: 'center'
-                        md_bg_color:1,0,0,1
-                        theme_text_color: "Custom"
-                        text_color: 1,1,1, 1
-                BoxLayout:
-                    orientation: 'vertical'
-                    size_hint_y: None
-                    height: dp(185)
-                    Button:
-                        id: ambulance_button
-                        size_hint: None, None
-                        size: (dp(290), dp(150))
-                        pos_hint: {'center_x': 0.5, 'center_y': 0.5}
-                        elevation:5
-                        md_bg_color: 1, 1, 1, 1
-                        on_release: root.animate_button('ambulance_button')
-                        Image:
-                            id:ambulance
-                            source: "images/ambulance.png"
-                            size_hint: None, None
-                            size:dp(290),dp(150)  # Set a default size
-                            pos:self.parent.pos
-                            allow_stretch: True
-                            keep_ratio: False
-                            md_bg_color: 1, 1, 1, 1
-                            padding: dp(10)
-                            size_hint_min: dp(290), dp(150)
-                    MDLabel:
-                        text: 'Ambulance Facility Manager'  # Description
-                        size_hint: None, None
-                        size: dp(290), dp(35)
-                        pos_hint: {'center_x': .5}
-                        font_size: "18sp"
-                        font_name: "Roboto"
-                        halign: 'center'
-                        md_bg_color:1,0,0,1
-                        theme_text_color: "Custom"
-                        text_color: 1,1,1,1
-                BoxLayout:
-                    orientation: 'vertical'
-                    size_hint_y: None
-                    height: dp(185)
-                    Button:
-                        id: gym_button
-                        size_hint: None, None
-                        size: (dp(290), dp(150))
-                        pos_hint: {'center_x': 0.5, 'center_y': 0.5}
-                        elevation:5
-                        md_bg_color: 1, 1, 1, 1
-                        on_release: root.animate_button('gym_button')  # Perform animation on button click
-                        Image:
-                            id:gym
-                            source: "images/gym.png"
-                            size_hint: None, None
-                            size:dp(290), dp(150)  # Set a default size
-                            pos:self.parent.pos
-                            allow_stretch: True
-                            keep_ratio: False
-                            md_bg_color: 1, 1, 1, 1
-                            padding: dp(10)
-                            size_hint_min: dp(290), dp(150)
-                    MDLabel:
-                        text: 'Gym Facility Manager'  # Description
-                        size_hint: None, None
-                        size: dp(290), dp(35)
-                        pos_hint: {'center_x': .5}
-                        font_size: "18sp"
-                        font_name: "Roboto"
-                        halign: 'center'
-                        md_bg_color:1,0,0,1
-                        theme_text_color: "Custom"
-                        text_color: 1,1,1,1
-#                BoxLayout:
-#                    orientation: 'vertical'
-#                    size_hint_y: None
-#                    height: dp(185)
-#                    MDIconButton:
-#                        id: doctors_button
-#                        size_hint: None, None
-#                        size: (dp(290), dp(150))
-#                        pos_hint: {'center_x': 0.5, 'center_y': 0.5}
-#                        elevation:5
-#                        md_bg_color: 1, 1, 1, 1
-#                        on_release: root.animate_button(self)
-#                        Image:
-#                            id:doctors
-#                            source: "images/doctor.png"
-#                            size_hint: None, None
-#                            size: self.parent.size if self.parent else (dp(290), dp(150))  # Set a default size
-#                            allow_stretch: True
-#                            keep_ratio: False
-#                            md_bg_color: 1, 1, 1, 1
-#                            padding: dp(10)
-#                            size_hint_min: dp(290), dp(150)  # Set a minimum size for the image
-#                    MDLabel:
-#                        text: 'Doctors'  # Description
-#                        size_hint: None, None
-#                        size: dp(290), dp(35)
-#                        pos_hint: {'center_x': .5}
-#                        font_size: 16
-#                        font_name: "MPoppins"
-#                        halign: 'center'
-#                        md_bg_color:1,1,1,1
-#                        theme_text_color: "Custom"
-#                        text_color: 0,0,0, 1
